rm spec issues re Object.create(null) & __proto__
the previous code used the idiom of `Object.create(null)` to use an object
like a hash or cache, and leveraged the prototype chain as a type of cache
miss failover. there is some js spec wierdness and changes around node 0.12
surrounding the existence of `__proto__` when using `Object.create(null)`.

to eliminate this confusion i've added a function to replace the above so
that the semantics are clear and consistent and we aren't using the
non-standard `__proto__`.